refactor (Patient): renamed PatientHome to CovidPatientHome and PatientHomeNonCovid to PatientHome
